--- a/Directory.Build.targets
+++ b/Directory.Build.targets
@@ -25,8 +25,8 @@
     FSharp.Compiler.Service: https://dev.azure.com/dnceng/internal/_release?view=mine&_a=releases&definitionId=98
 
     -->
-    <PackageReference Update="FSharp.Core" Version="6.0.4" />
-    <PackageReference Update="FSharp.Compiler.Service" Version="41.0.4" />
+    <PackageReference Update="FSharp.Core" Version="6.0.5-beta.22327.2" />
+    <PackageReference Update="FSharp.Compiler.Service" Version="41.0.5-preview.22327.2" />
   </ItemGroup>
 
 </Project>