## Before this can be pulled out of draft

- [x] Publish F# packages from the internal build [20220627.2](https://dev.azure.com/dnceng/internal/_build/results?buildId=1848427&view=results) from the `Packages/Shipping` artifact, version `*.22327.2`.

**tl;dr** - Initial package restore unchanged, 4.2s.  Cached package restore 1.5s.  Subsequent restores within the same process (i.e., no longer JITing anything): 0.1s.

### Long Version

Some local tests showed:

- With old NuGet package manager:
  1. Create new blank notebook.
  2. Execute `1+1` to start the backing process and warm up the kernel.
  3. Execute `#r "nuget:SomePackage, 1.0.0"`
  4. Result: 4.2s

- With new NuGet package manager, cache not populated:
  1. Create new blank notebook.
  2. Execute `1+1` to start the backing process and warm up the kernel.
  3. Execute `#r "nuget:SomePackage, 1.0.0"`
  4. Result: 4.2s, and cache created on disk.

- With new NuGet package manager, cache populated by previous run:
  1. Create new blank notebook.
  2. Execute `1+1` to start the backing process and warm up the kernel.
  3. Execute `#r "nuget:SomePackage, 1.0.0"`
  4. Result: 1.5s, cache used from disk, package manager was jitted.
  5. Execute `#r "nuget:SomePackage, 1.0.0"` again
  6. Result: 0.1s, cache used from disk, package manager native code was re-used.